requirements.txt
# Worklog Manager - Python Dependencies
# Version 1.7.0 - Advanced Features

# Core Dependencies (included with Python standard library):
# - tkinter (GUI framework)
# - sqlite3 (database)
# - datetime (time handling)
# - threading (background timers)
# - logging (application logging)
# - json (data export/settings)
# - csv (data export)
# - uuid (unique identifiers)
# - pathlib (file path handling)
# - atexit (cleanup handlers)

# Required Dependencies for Advanced Features:
# Install with: pip install -r requirements.txt

# Cross-platform notifications (Phase 4)
plyer>=2.0

# PDF export functionality (Phase 3 & 4)
reportlab>=3.6.0

# System tray integration and icon rendering (Windows/Linux)
pystray>=0.19.5
pillow>=10.0.0

# Native Windows integrations required by pystray
pywin32>=305; platform_system == "Windows"

# Python 3.6 compatibility
dataclasses; python_version < "3.7"

# Optional Dependencies for Enhanced Experience:

# Packaging / distribution helpers
# pyinstaller>=6.3.0

# Development Dependencies (optional):
# Uncomment these for development/testing

# Code formatting
# black>=22.0.0

# Code linting  
# flake8>=4.0.0

# Type checking
# mypy>=0.950

# Testing framework
# pytest>=7.0.0

# Installation Commands:
# 
# Basic installation (Phases 1-3 features):
#   pip install reportlab
#
# Full installation (all Phase 4 features):
#   pip install plyer reportlab
#
# Windows full installation:
#   pip install plyer reportlab pywin32
#
# Development setup:
#   pip install plyer reportlab black flake8 mypy pytest
#
# Note: The application gracefully handles missing optional dependencies
# and will inform users about unavailable features.